+++ stocks-dataset/src/main/resources/schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE stocks (
 	id int not null auto_increment,
     symbol varchar(32) not null,
     price float not null,
-    volume int not null,
+    volume bigint not null,
     date date,
     primary key (id)
 );
